+++ database/src/main/resources/db/migration/V1_48__Add_cache_update_for_chain_versions.sql
@@ -1,0 +1,2 @@
+INSERT INTO cache_update(cache_key, description, cache_value, last_updated)
+VALUES ('chain_releases', 'Set of all chain releases', null, now());